+++ include/et/rendering/interface/buffer.h
@@ -40,8 +40,8 @@ public:
 
 	struct Description
 	{
-		uint32_t size = 0;
-		uint32_t alignedSize = 0;
+		uint64_t size = 0;
+		uint64_t alignedSize = 0;
 		Usage usage = Usage::Constant;
 		Location location = Location::Device;
 		BinaryDataStorage initialData;
@@ -51,16 +51,16 @@ public:
 	Buffer() = default;
 	virtual ~Buffer() = default;
 
-	virtual uint8_t* map(uint32_t begin, uint32_t length) = 0;
+	virtual uint8_t* map(uint64_t begin, uint64_t length) = 0;
 	virtual void modifyRange(uint64_t begin, uint64_t length) = 0;
 	virtual void unmap() = 0;
 
 	virtual bool mapped() const = 0;
 
-	virtual void updateData(uint32_t offset, const BinaryDataStorage&) = 0;
-	virtual void transferData(Buffer::Pointer destination, uint32_t srcOffset, uint32_t dstOffset, uint32_t size) = 0;
+	virtual void updateData(uint64_t offset, const BinaryDataStorage&) = 0;
+	virtual void transferData(Buffer::Pointer destination, uint64_t srcOffset, uint64_t dstOffset, uint64_t size) = 0;
 
-	virtual uint32_t size() const = 0;
+	virtual uint64_t size() const = 0;
 };
 
 }
